RTL, Arabic, Tinymce, Block Outcomes
A few changes! Arabic has been added as a development language, which
require some changes to support RTL. These are experimental and may
need review. Tinymce has received a minor update to fix bugs. And the
Unit Planner now has the beginnings of outcomes in blocks. This needs
expanding to work in deployed units, and to carry through to lesson
outcomes. Coming soon.