System Admin: apply translation function to timezones in System Admin (#478)
Hi Sandra, I appreciate your honesty here, and am not sure that there is a right answer. I certainly prefer the new drop down approach over the old link+copy/paste. The way I see things, if people don't translate these new strings it is not the end of the world, as they appear in English and that is fine in a lot of cases. However, for those who do want 100% translation, then that option is available. Let's keep this in for v16 for now and see what happens...we can always back out down the line. Thanks for getting on your soapbox ; ) Ross.