Name clarification and pre-checked notifications
Just a couple of minor changes prompted by feedback from professors. Namely, changing the label of the "Name" field to "Name of Lesson Plan". It seems there was some confusion as to what name to put in there! Also, the "Notify all class participants check box" is now checked by default. Some professors were forgetting to check and it seemed that an opt-out approach was less hazardous.